a base implementation for one-session-one-run mode

+++ mriqc/interfaces/qc.py
@@ -7,7 +7,7 @@
 # @Date:   2016-01-05 11:29:40
 # @Email:  code@oscaresteban.es
 # @Last modified by:   oesteban
-# @Last Modified time: 2016-04-15 15:39:39
+# @Last Modified time: 2016-04-20 09:12:50
 """ Nipype interfaces to quality control measures """
 
 import numpy as np
@@ -94,7 +94,7 @@ class StructuralQC(BaseInterface):
             snrvals.append(snr(inudata, segdata, airdata, fglabel=tlabel,
                                erode=erode))
             self._results['snr'][tlabel] = snrvals[-1]
-        self._results['snr']['total'] = np.mean(snrvals)
+        self._results['snr']['total'] = float(np.mean(snrvals))
 
         # CNR
         self._results['cnr'] = cnr(inudata, segdata)
@@ -128,27 +128,28 @@ class StructuralQC(BaseInterface):
                                     'p95': p95, 'p05': p05}
 
         # Image specs
-        self._results['size'] = {'x': imdata.shape[0],
-                                 'y': imdata.shape[1],
-                                 'z': imdata.shape[2]}
+        self._results['size'] = {'x': int(imdata.shape[0]),
+                                 'y': int(imdata.shape[1]),
+                                 'z': int(imdata.shape[2])}
         self._results['spacing'] = {
-            i: v for i, v in zip(['x', 'y', 'z'],
-                                 imnii.get_header().get_zooms()[:3])}
+            i: float(v) for i, v in zip(
+                ['x', 'y', 'z'], imnii.get_header().get_zooms()[:3])}
 
         try:
-            self._results['size']['t'] = imdata.shape[3]
+            self._results['size']['t'] = int(imdata.shape[3])
         except IndexError:
             pass
 
         try:
-            self._results['spacing']['tr'] = imnii.get_header().get_zooms()[3]
+            self._results['spacing']['tr'] = float(imnii.get_header().get_zooms()[3])
         except IndexError:
             pass
 
         # Bias
         bias = nb.load(self.inputs.in_bias).get_data()[segdata > 0]
         self._results['inu'] = {
-            'range': np.abs(np.percentile(bias, 95.) - np.percentile(bias, 5.)), 'med': np.median(bias)}  #pylint: disable=E1101
+            'range': float(np.abs(np.percentile(bias, 95.) - np.percentile(bias, 5.))),
+            'med': float(np.median(bias))}  #pylint: disable=E1101
 
 
         # Flatten the dictionary
